LayoutTests/accessibility/ios-simulator/attributed-string-for-range-expected.txt

hello     world
This test ensures that attributed string for range works

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Attributed string: hello  {
    UIAccessibilityTokenFontFamily = "Times New Roman";
    UIAccessibilityTokenFontName = "Times New Roman";
    UIAccessibilityTokenFontSize = 16;
}￼{
    UIAccessibilityTokenAttachment = "WebAccessibilityObjectWrapper: cake";
}  {
    UIAccessibilityTokenFontFamily = "Times New Roman";
    UIAccessibilityTokenFontName = "Times New Roman";
    UIAccessibilityTokenFontSize = 16;
}￼{
    UIAccessibilityTokenAttachment = "WebAccessibilityObjectWrapper: action, title, subtitle";
} {
    UIAccessibilityTokenFontFamily = "Times New Roman";
    UIAccessibilityTokenFontName = "Times New Roman";
    UIAccessibilityTokenFontSize = 16;
}
PASS successfullyParsed is true

TEST COMPLETE
